Revert "TUDO PRONTO"
This reverts commit ecf17391c158a3df0eaf8c5772eb6257e3d4b1ad.

+++ projeto/src/main/java/br/com/fiap/projeto/controller/ClienteResource.java
@@ -54,7 +54,7 @@ public class ClienteResource implements ContainerResponseFilter {
 	
 	@DELETE
 	@Path("/{cpf}")
-	public Response delete(@PathParam("cpf") long cpf) {
+	public Response delete(@PathParam("cpf") int cpf) {
 		if (ClienteRepository.delete(cpf)) {
 			ResponseBuilder response = Response.noContent();
 			return response.build();
